Improve Trade Service
Add an isProducer property to CityResourcesServices as a resource have
to have both Deposit on the map AND active production building.
Added Previous Era Ressource when trading to be able to trade current
era resources with not produced resources from previous era when
there’s remaining current era resources available.